Organize code into respective packages
