+++ src/main/java/com/duke/Duke.java
@@ -1,4 +1,8 @@
-package com.iP.yiheng;
+package com.duke;
+
+import com.duke.parser.Parser;
+import com.duke.task.TaskList;
+import com.duke.ui.UserInterface;
 
 import java.util.Locale;
 import java.util.Scanner;
