+++ text-ui-test/EXPECTED.TXT
@@ -1,25 +1,25 @@
-Duke: Hello! I'm Duke
+com.duke.Duke: Hello! I'm com.duke.Duke
 What can I do for you?
 
-Duke: Got it. I have added this task:
+com.duke.Duke: Got it. I have added this task:
 [T][ ]  say hello
 Now you have 1 tasks in the list.
 
-Duke: Got it. I have added this task:
+com.duke.Duke: Got it. I have added this task:
 [D][ ]  project (by: sunday)
 Now you have 2 tasks in the list.
 
-Duke: Got it. I have added this task:
+com.duke.Duke: Got it. I have added this task:
 [E][ ]  splashdown (at: nus)
 Now you have 3 tasks in the list.
 
-Duke: Nice! I've marked this task as done:
+com.duke.Duke: Nice! I've marked this task as done:
 [T][X]  say hello
 
-Duke: Nice! I've marked this task as done:
+com.duke.Duke: Nice! I've marked this task as done:
 [D][X]  project (by: sunday)
 
-Duke: Noted. I've removed this task:
+com.duke.Duke: Noted. I've removed this task:
 [E][ ]  splashdown (at: nus)
 Now you have 2 tasks in the list.
 
@@ -28,4 +28,4 @@ Here are the tasks in your list:
 1. [T][X]  say hello
 2. [D][X]  project (by: sunday)
 
-Duke: Bye. Hope to see you again soon!
+com.duke.Duke: Bye. Hope to see you again soon!
